Move non-test classes to mocks folder
