+++ src/EPR.SubsidiaryBulkUpload.Application.UnitTests/Mocks/TestCsvReaderClassParamMap.cs
@@ -1,6 +1,6 @@
 ﻿using CsvHelper.Configuration;
 
-namespace EPR.SubsidiaryBulkUpload.Application.UnitTests.Services;
+namespace EPR.SubsidiaryBulkUpload.Application.UnitTests.Mocks;
 
 public class TestCsvReaderClassParamMap : ClassMap<TestCsvReaderParamClass>
 {
