+++ src/EPR.SubsidiaryBulkUpload.Application.UnitTests/Services/CustomCsvReaderTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text;
 using CsvHelper;
 using EPR.SubsidiaryBulkUpload.Application.CsvReaderConfiguration;
+using EPR.SubsidiaryBulkUpload.Application.UnitTests.Mocks;
 
 namespace EPR.SubsidiaryBulkUpload.Application.UnitTests.Services;
 
@@ -29,7 +30,7 @@ public class CustomCsvReaderTests
         using var reader = new StreamReader(stream);
         var map = new TestCsvReaderClassParamMap();
         var invalidHeaders = new List<InvalidHeader>(); // Initialize with appropriate values
-        var customCsvReader = new CustomCsvReaderTest(reader, configuration);
+        var customCsvReader = new CustomCsvReaderWrapper(reader, configuration);
 
         customCsvReader.Context.RegisterClassMap<TestCsvReaderClassParamMap>();
         customCsvReader.Read();
